requirements.txt
pillow
requests
pyrebase
flask
cloudinary
sendgrid
gunicorn
